[refactor/#135] UserStat refactor
